crisp_backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
stripe==12.3.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
dj-database-url==2.3.0
python-dateutil==2.9.0.post0